requirements.txt
streamlit==1.32.0
pandas==2.2.1
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
youtube-transcript-api==0.6.2
openai==1.12.0
python-dotenv==1.0.1
isodate==0.6.1
plotly==5.19.0 